+++ src/lib/viz/popups/Popup.ts
@@ -1,5 +1,4 @@
 import { Deck } from '@deck.gl/core';
-import { format as d3Format } from 'd3-format';
 import { CartoPopupError, popupErrorTypes } from '../errors/popup-error';
 import { getMapContainer } from '../utils/map-utils';
 
@@ -58,7 +57,6 @@ export class Popup {
         }
       }
     });
-
     this._render();
   }
 
@@ -74,7 +72,7 @@ export class Popup {
 
     this._coordinates = coordinates;
 
-    if (this._deckInstance && this._isOpen) {
+    if (this._deckInstance) {
       this._render();
     }
   }
@@ -115,6 +113,7 @@ export class Popup {
     }
 
     this._isOpen = true;
+    this._render();
   }
 
   /**
@@ -179,7 +178,12 @@ export class Popup {
   }
 
   private _render() {
-    if (this._coordinates && this.getContent().trim().length > 0 && this._deckInstance) {
+    if (
+      this._isOpen &&
+      this._coordinates &&
+      this.getContent().trim().length > 0 &&
+      this._deckInstance
+    ) {
       const pixels = coordinates2pixels(this._coordinates, this._deckInstance);
 
       if (pixels) {
@@ -265,21 +269,15 @@ function generatePopupContent(elements: any, features: Record<string, any>[]): s
 
           let elementValue = feature.properties[attr];
 
-          if (format && typeof format === 'function') {
-            elementValue = format(elementValue);
-          } else if (format && typeof format === 'string') {
-            let formatter;
-
-            try {
-              formatter = d3Format(format);
-            } catch (err) {
+          if (format) {
+            if (typeof format === 'function') {
+              elementValue = format(elementValue);
+            } else {
               throw new CartoPopupError(
-                `The format '${format}' is not a recognized D3 format`,
+                `Invalid popup format: '${format}' is not a function`,
                 popupErrorTypes.FORMAT_INVALID
               );
             }
-
-            elementValue = formatter(elementValue);
           }
 
           return `<p class="as-body">${title}</p>
@@ -305,13 +303,10 @@ export interface PopupElement {
   title?: string | null;
 
   /**
-   * d3 format for the value of this attribute.
+   * Format function
    */
-  format?: string | FormatFunction;
+  format?: (value: any) => any | null;
 }
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type FormatFunction = (value: any) => any;
 
 /**
  * Popup options
@@ -351,17 +346,6 @@ interface PopupOptions {
     | 'bottom-left'
     | 'bottom-right';
 }
-
-// function pixels2coordinates(pixels: number[], deckInstance?: Deck) {
-//   let coordinates;
-
-//   if (deckInstance) {
-//     const viewport = deckInstance.getViewports(undefined)[0];
-//     coordinates = viewport.unproject(pixels);
-//   }
-
-//   return coordinates;
-// }
 
 function coordinates2pixels(coordinates: number[], deckInstance?: Deck) {
   let pixels;